3/16 Update
In this commit, I split the "media" items into a separate template so that it does not generate all items on every device. VMware virtual machine discovery will create the device in monitoring with the collective template, you will need to manually assign the media template to the device.

Items added/changed:
VMware Datastore: Added free space (% and B), Free space (% and B), total size.

Radarr health check: Checks Radarr API to read for the 'message' string. If the string does not exist, it reports as healthy. If the 'message' string exists, it will trigger and display the message as the operational data.

Sonarr health check: Checks Sonarr API to read for the 'message' string. If the string does not exist, it reports as healthy. If the 'message' string exists, it will trigger and display the message as the operational data.

CPU triggers: Corrected the "resolve" expression from "<=" to "<". This was causing flapping of alerts on CPU if the number read was equal to.